+++ src/util/ranges2diagnostics.ts
@@ -4,12 +4,14 @@ import { getCodeWithURI } from './getCodeWithURI';
 
 export default function ranges2diagnostics(
     code: LLCode,
-    message: string,
+    messages: string | string[],
     ranges: vscode.Range[]
 ): vscode.Diagnostic[] {
-    return ranges.map(range => ({
+    if (typeof messages === 'string') messages = [messages];
+    else console.assert(messages.length === ranges.length);
+    return ranges.map((range, i) => ({
         code: getCodeWithURI(code),
-        message: message,
+        message: messages[messages.length === 1 ? 0 : i],
         range: range,
         severity: severity[code],
         source: extensionDisplayName,